+++ examples/src/example-flow-field-pathfinding.ts
@@ -9,15 +9,14 @@ import {
     getNodeRefType,
     NodeType,
 } from 'navcat';
+import { generateTiledNavMesh, type TiledNavMeshInput, type TiledNavMeshOptions } from 'navcat/blocks';
+import { createNavMeshHelper, createNavMeshPolyHelper, type DebugObject, getPositionsAndIndices } from 'navcat/three';
 import * as THREE from 'three';
 import { LineGeometry, OrbitControls } from 'three/examples/jsm/Addons.js';
 import { Line2 } from 'three/examples/jsm/lines/webgpu/Line2.js';
 import { Line2NodeMaterial } from 'three/webgpu';
-import { createNavMeshHelper, createNavMeshPolyHelper, type DebugObject } from './common/debug';
 import { createExample } from './common/example-base';
 import { computeUniformCostFlowField, type FlowField, getNodePathFromFlowField } from './common/flow-field';
-import { generateTiledNavMesh, type TiledNavMeshInput, type TiledNavMeshOptions } from './common/generate-tiled-nav-mesh';
-import { getPositionsAndIndices } from './common/get-positions-and-indices';
 import { loadGLTF } from './common/load-gltf';
 
 /* setup example scene */