+++ examples/src/example-tiled-navmesh.ts
@@ -1,6 +1,5 @@
 import GUI from 'lil-gui';
-import * as THREE from 'three';
-import { OrbitControls } from 'three/examples/jsm/Addons.js';
+import { generateTiledNavMesh, type TiledNavMeshInput, type TiledNavMeshOptions } from 'navcat/blocks';
 import {
     createCompactHeightfieldDistancesHelper,
     createCompactHeightfieldRegionsHelper,
@@ -15,11 +14,12 @@ import {
     createRawContoursHelper,
     createSimplifiedContoursHelper,
     createTriangleAreaIdsHelper,
+    getPositionsAndIndices,
     type DebugObject,
-} from './common/debug';
+} from 'navcat/three';
+import * as THREE from 'three';
+import { OrbitControls } from 'three/examples/jsm/Addons.js';
 import { createExample } from './common/example-base';
-import { generateTiledNavMesh, type TiledNavMeshInput, type TiledNavMeshOptions } from './common/generate-tiled-nav-mesh';
-import { getPositionsAndIndices } from './common/get-positions-and-indices';
 import { loadGLTF } from './common/load-gltf';
 
 /* setup example scene */