+++ src/index.ts
@@ -1,4 +1,5 @@
-export * from "./debug";
+export type { Box3, Vec3 } from 'maaths';
+export * from './debug';
 export * from './generate';
 export * as geometry from './geometry';
 export * from './query';